moving book dto into its own file
