+++ apis/src/main/java/xyz/bookself/controllers/user/BookIdListIdDTO.java
@@ -1,0 +1,22 @@
+package xyz.bookself.controllers.user;
+
+public class BookIdListIdDTO {
+    private String bookId;
+    private String listId;
+
+    public String getBookId() {
+        return bookId;
+    }
+
+    public String getListId() {
+        return listId;
+    }
+
+    public void setBookId(String bookId) {
+        this.bookId = bookId;
+    }
+
+    public void setListId(String listId) {
+        this.listId = listId;
+    }
+}
