+++ apis/src/main/java/xyz/bookself/controllers/user/BookListController.java
@@ -52,24 +52,3 @@ public class BookListController {
         return new ResponseEntity<>(foundBookList, HttpStatus.OK);
     }
 }
-
-class BookIdListIdDTO {
-    private String bookId;
-    private String listId;
-
-    public String getBookId() {
-        return bookId;
-    }
-
-    public String getListId() {
-        return listId;
-    }
-
-    public void setBookId(String bookId) {
-        this.bookId = bookId;
-    }
-
-    public void setListId(String listId) {
-        this.listId = listId;
-    }
-}
